fix: 예약 시간 추가하도록 테스트 수정
