--- src/main/resources/data.sql
@@ -1,3 +1,0 @@
-insert into reservation_time (id, start_at)
-VALUES (1, '10:00'),
-       (2, '11:00');
